Move signup interfaces to common
