--- packages/amplify-ui-components/src/components/amplify-sign-up/amplify-sign-up-interface.ts
@@ -1,14 +1,0 @@
-export interface AmplifySignUpFormFooterProps {
-  submitButtonText: string;
-  haveAcccountText: string;
-  signInText: string;
-  handleAuthStateChange?: any;
-}
-
-export interface AmplifySignUpAttributes {
-  username: string;
-  password: string;
-  attributes?: {
-    [userAttributes: string]: string;
-  };
-}
